Add Spark Standalone cluster with README
Add the components required for a local Spark Standalone cluster housed
within Docker containers. In addition, add scaffolding required for a
test `SparkPi` job.

Lastly, start assembling a `README` to document specifics about the
Spark environment so that we have a concise reference as we move toward
the EC2 based deployment.